Add retrofit Query

+++ src/main/java/ex/com/bakingapp/data/api/Ingredient.java
@@ -3,16 +3,17 @@ package ex.com.bakingapp.data.api;
 
 public class Ingredient {
 
-    private Integer quantity;
+    private float quantity;
     private String measure;
     private String ingredient;
 
-    public Integer getQuantity() {return quantity;}
-    public void setQuantity(Integer quantity) {this.quantity = quantity;}
+    public float getQuantity() {return quantity;}
+    public void setQuantity(float quantity) {this.quantity = quantity;}
 
     public String getMeasure() {return measure;}
     public void setMeasure(String measure) {this.measure = measure;}
 
     public String getIngredient() {return ingredient;}
     public void setIngredient(String ingredient) {this.ingredient = ingredient;}
+
 }
